Original Ego4d testing working for Action Recognition.
Eval script for action recognition is bugged in Ego4d. In LTA 'is_last_clip' is always true, hence the end_time counter is never used (always reset on is_last_clip=True). However, this is NOT the case for Action Recognition, hence the end-time builds up for new clips, but these clips are sampled for NEW annotation's time-ranges! Therefore this goes out of bounds and videos can't be sampled.

Instead with uniform sampling we always take the first segment of a clip-annotation range.